refactor: domain packaging 변경
- domain/model
- domain/usecase
