+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/pomodoro/setting/PomodoroViewModel.kt
@@ -2,8 +2,8 @@ package com.pomonyang.mohanyang.presentation.screen.pomodoro.setting
 
 import androidx.lifecycle.viewModelScope
 import com.pomonyang.mohanyang.data.repository.PomodoroSettingRepository
-import com.pomonyang.mohanyang.domain.GetPomodoroSettingListUseCase
-import com.pomonyang.mohanyang.domain.PomodoroCategoryModel
+import com.pomonyang.mohanyang.domain.model.PomodoroCategoryModel
+import com.pomonyang.mohanyang.domain.usecase.GetPomodoroSettingListUseCase
 import com.pomonyang.mohanyang.presentation.base.BaseViewModel
 import com.pomonyang.mohanyang.presentation.base.ViewEvent
 import com.pomonyang.mohanyang.presentation.base.ViewSideEffect
